Moves the watcher informer factory creation to a function
