+++ cmd/nsdns/utils.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"os"
 	"path"
+	"time"
 )
 
 import (
@@ -16,10 +17,13 @@ import (
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/client-go/util/homedir"
+	"k8s.io/client-go/informers"
+	"k8s.io/client-go/tools/cache"
 )
 
 import (
 	"github.com/Eagerod/kube-namesilo-dns/pkg/namesilo_api"
+	"github.com/Eagerod/kube-namesilo-dns/pkg/nsdns"
 )
 
 type RecordReconciliation struct {
@@ -65,4 +69,36 @@ func GetKubernetesClientSet() (*kubernetes.Clientset, error) {
 	}
 
 	return nil, errors.New("failed to configure Kubernetes client")
+}
+
+func DomainManagerInformerFactory(dm *nsdns.DnsManager, clientset *kubernetes.Clientset) informers.SharedInformerFactory {
+	informerFactory := informers.NewSharedInformerFactory(clientset, time.Minute)
+
+	informerFactory.Networking().V1().Ingresses().Informer().AddEventHandler(
+		cache.ResourceEventHandlerFuncs{
+			AddFunc: func(obj interface{}) {
+				ingress := obj.(*apinetworkingv1.Ingress)
+
+				if err := dm.HandleIngressExists(ingress); err != nil {
+					log.Error(err)
+				}
+			},
+			DeleteFunc: func(obj interface{}) {
+				ingress := obj.(*apinetworkingv1.Ingress)
+
+				if err := dm.HandleIngressDeleted(ingress); err != nil {
+					log.Error(err)
+				}
+			},
+			UpdateFunc: func(old, new interface{}) {
+				ingress := new.(*apinetworkingv1.Ingress)
+
+				if err := dm.HandleIngressExists(ingress); err != nil {
+					log.Error(err)
+				}
+			},
+		},
+	)
+
+	return informerFactory
 }
