+++ cmd/nsdns/watch.go
@@ -10,9 +10,6 @@ import (
 import (
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	networkingv1 "k8s.io/api/networking/v1"
-	"k8s.io/client-go/informers"
-	"k8s.io/client-go/tools/cache"
 )
 
 import (
@@ -52,33 +49,7 @@ func watchCommand() *cobra.Command {
 				return err
 			}
 
-			informerFactory := informers.NewSharedInformerFactory(clientset, time.Minute)
-
-			informerFactory.Networking().V1().Ingresses().Informer().AddEventHandler(
-				cache.ResourceEventHandlerFuncs{
-					AddFunc: func(obj interface{}) {
-						ingress := obj.(*networkingv1.Ingress)
-
-						if err := dm.HandleIngressExists(ingress); err != nil {
-							log.Error(err)
-						}
-					},
-					DeleteFunc: func(obj interface{}) {
-						ingress := obj.(*networkingv1.Ingress)
-
-						if err := dm.HandleIngressDeleted(ingress); err != nil {
-							log.Error(err)
-						}
-					},
-					UpdateFunc: func(old, new interface{}) {
-						ingress := new.(*networkingv1.Ingress)
-
-						if err := dm.HandleIngressExists(ingress); err != nil {
-							log.Error(err)
-						}
-					},
-				},
-			)
+			informerFactory := DomainManagerInformerFactory(dm, clientset)
 
 			stop := make(chan struct{})
 			informerFactory.Start(stop)
